requirements.txt
fastapi[all]
pydantic
pydantic-settings
pyyaml
pylint
pytest-env

cp-mgmt-api-sdk
